WIP:Fixing srun mode

+++ scoop/launch/workerLaunch.py
@@ -30,7 +30,7 @@ class Host(object):
        Can be remote (ssh via netowrk) or represent localhost."""
     BOOTSTRAP_MODULE = 'scoop.bootstrap.__main__'
     BASE_SSH = ['ssh', '-x', '-n', '-oStrictHostKeyChecking=no']
-    BASE_SRUN = ['srun','--exclusive','--ntasks=1','bash','-c']
+    BASE_SRUN = ['srun','--exclusive','--ntasks=1','--nodes=1','bash','-c']
 
     LAUNCHING_ARGUMENTS = namedtuple(
         'launchingArguments',
@@ -268,7 +268,7 @@ class Host(object):
         if not self.isLocal() and self.remoteProcessGID is None:
                 scoop.logger.warn("Zombie process(es) possibly left on "
                              "host {0}!".format(self.hostname))
-        elif not self.isLocal():
+        elif not self.isLocal() and not self.use_srun:
             command = ("python -c "
                        "'import os, signal; os.killpg({0}, signal.SIGKILL)' "
                        ">&/dev/null").format(self.remoteProcessGID)
